Moved c/cpp files to lib
