+++ lib/imgui/imgui_impl_opengl3.cpp
@@ -117,7 +117,7 @@
 #elif defined(IMGUI_IMPL_OPENGL_LOADER_GLEW)
 #include <GL/glew.h>    // Needs to be initialized with glewInit() in user's code
 #elif defined(IMGUI_IMPL_OPENGL_LOADER_GLAD)
-#include <glad.h>  // Needs to be initialized with gladLoadGL() in user's code
+#include <glad/glad.h>  // Needs to be initialized with gladLoadGL() in user's code
 #elif defined(IMGUI_IMPL_OPENGL_LOADER_GLBINDING)
 #include <glbinding/gl/gl.h>  // Initialize with glbinding::initialize()
 #include <glbinding/glbinding.h>
