+++ lib/imgui/imgui_impl_opengl3.h
@@ -51,7 +51,7 @@ IMGUI_IMPL_API void     ImGui_ImplOpenGL3_DestroyDeviceObjects();
     #if defined(__has_include)
         #if __has_include(<GL/glew.h>)
             #define IMGUI_IMPL_OPENGL_LOADER_GLEW
-        #elif __has_include(<glad.h>)
+        #elif __has_include(<glad/glad.h>)
             #define IMGUI_IMPL_OPENGL_LOADER_GLAD
         #elif __has_include(<GL/gl3w.h>)
             #define IMGUI_IMPL_OPENGL_LOADER_GL3W
