+++ src/Diffusion.h
@@ -22,7 +22,7 @@
 #include "BEM.h"
 #include "GuiHelper.h"
 
-#include <json.hpp>
+#include <json/json.hpp>
 
 #include <cstdlib>
 #include <iostream>
