+++ src/FlowFeature.h
@@ -11,7 +11,7 @@
 #include "Body.h"
 #include "Feature.h"
 #include "ElementPacket.h"
-#include <json.hpp>
+#include <json/json.hpp>
 
 #include <memory>
 #include <iostream>
