+++ src/MeasureFeature.h
@@ -12,7 +12,7 @@
 #include "Omega2D.h"
 #include "Feature.h"
 #include "ElementPacket.h"
-#include <json.hpp>
+#include <json/json.hpp>
 
 #include <iostream>
 #include <vector>
