+++ src/OglHelper.h
@@ -17,7 +17,7 @@
   #define GL_APIENTRY_DEFINED
 #endif // APIENTRY
 
-#include "glad.h"
+#include <glad/glad.h>
 
 #include <cstdint>
