+++ src/VtkXmlWriter.h
@@ -16,8 +16,8 @@
 #include <string>
 #include <vector>
 
-#include <base64_rfc4648.hpp>
-#include "tinyxml2.h"
+#include <cppcodec/base64_rfc4648.hpp>
+#include <tinyxml2/tinyxml2.h>
 
 class VtkXmlWriter {
 public:
